DD-9 refactor global로 사용하는 response 폴더 위치 dto로 변경
